pyproject.toml
[tool.poetry]
name = "gve-devnet-meraki-snapshot-collector"
version = "0.1.0"
description = "Utility to collect Meraki MV camera snapshots for a given time"
authors = ["Matt Schmitz <mattsc@cisco.com>"]
license = "CISCO SAMPLE CODE LICENSE"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
typer = "^0.9.0"
rich = "^13.7.0"
meraki = "^1.42.0"
python-dotenv = "^1.0.1"
jinja2 = "^3.1.3"

[tool.poetry.group.dev.dependencies]
ruff = "^0.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
